cifs: find and use the dentry for cached non-root directories also
BugLink: https://bugs.launchpad.net/bugs/2036450

This allows us to use cached attributes for the entries in a cached
directory for as long as a lease is held on the directory itself.
Previously we have always allowed "used cached attributes for 1 second"
but this extends this to the lifetime of the lease as well as making the
caching safer.

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
(cherry picked from commit e4029e072673d8a694f660f551609dd4f9265088)
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>