cifs: allow dumping keys for directories too
BugLink: https://bugs.launchpad.net/bugs/2036450

Dumping the enc/dec keys is a session wide operation.
And it should not matter if the ioctl was run on
a regular file or a directory.

Currently, we obtain the tcon pointer from the
cifs file handle. But since there's no dir open call
in cifs, this is not populated for dirs.

This change allows dumping of session keys using ioctl
even for directories. To do this, we'll now get the
tcon pointer from the superblock, and not from the file
handle.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
(cherry picked from commit b3edef6b9cd0b8893a5649ab1fa8bce71045d139)
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>